Apply the Postgres bootstrap to a real server, and fix what that found
The offline checks proved the model was expressible and that the two bootstraps
agreed with each other. Neither executes a statement, so neither could catch
the one thing that actually broke.

PgSchemaCheck now takes an optional EVECONSOLE_PG connection string and does the
real thing: EnsureCreated, then PostgresSchema, then counts what landed, checks
each seed inserted its rows, and applies a second time to prove that a normal
second launch is idempotent. Taken from the environment rather than an argument
so a password does not end up in shell history or a CI log.

Running it found the bug:

  42804: column "StationFilter" is of type bigint but expression is of type text

In an INSERT ... SELECT ... UNION ALL, PostgreSQL settles the union's column
types before it looks at the target, and an untyped NULL settles as text.
SQLite is dynamically typed and never had an opinion, so the statement it was
copied from was fine. NULL::bigint fixes it. No amount of reading the file
would have shown this.

Verified against PostgreSQL 18.6 as a role that is neither superuser nor
createdb, owning only its own database: 199 tables, 254 indexes, all six seeds
at their expected row counts, second apply idempotent.

Worth recording from that run: an owner of a database still gets CREATE on
public under PG15+ rules, which is what makes the least-privilege setup
(CREATE DATABASE x OWNER appuser) work without any further grants.

+++ Data/PostgresSchema.cs
@@ -128,6 +128,13 @@ public static class PostgresSchema
     /// <para>The <c>WHERE NOT EXISTS</c> forms are left exactly as they are: they are already
     /// portable, and they mean "seed only an empty table", which is not the same thing as
     /// per-row conflict handling and must not be rewritten into it.</para>
+    ///
+    /// <para>⚠️ <c>NULL::bigint</c>, not a bare <c>NULL</c>. In an <c>INSERT … SELECT … UNION
+    /// ALL</c> PostgreSQL settles the union's column types before it ever looks at the target,
+    /// and an untyped NULL settles as <c>text</c> — so the insert fails with "column
+    /// StationFilter is of type bigint but expression is of type text". SQLite is dynamically
+    /// typed and never had an opinion. This was caught by applying the file to a real server;
+    /// no amount of reading it would have shown it.</para>
     /// </summary>
     private static readonly string[] Seeds =
     [
@@ -140,10 +147,10 @@ public static class PostgresSchema
         """
         INSERT INTO "MarketPricingConfigs"
             ("Method", "LocationName", "LocationId", "PriceType", "IsEnabled", "SortOrder", "LastStatus", "StationFilter", "UsePercentileFilter", "PercentilePercent")
-        SELECT 'Region', 'The Forge', 10000002, 'Midpoint', true, 0, '', NULL, true, 1.0
+        SELECT 'Region', 'The Forge', 10000002, 'Midpoint', true, 0, '', NULL::bigint, true, 1.0
         WHERE NOT EXISTS (SELECT 1 FROM "MarketPricingConfigs")
         UNION ALL
-        SELECT 'Region', 'Domain',    10000043, 'Midpoint', true, 1, '', NULL, true, 1.0
+        SELECT 'Region', 'Domain',    10000043, 'Midpoint', true, 1, '', NULL::bigint, true, 1.0
         WHERE NOT EXISTS (SELECT 1 FROM "MarketPricingConfigs")
         """,
         """
